requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Vector Databases and RAG
faiss-cpu>=1.7.4
chromadb>=0.4.0
pinecone-client>=2.2.0
weaviate-client>=3.20.0

# LLM and Agent Frameworks
langchain>=0.1.0
langchain-community>=0.0.20
llama-index>=0.9.0
autogen-agentchat>=0.2.0
openai>=1.0.0
anthropic>=0.3.0

# Audio Processing
librosa>=0.10.0
whisper-openai
bark

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0

# Web and API
streamlit>=1.28.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Model Optimization
bitsandbytes>=0.41.0
peft>=0.5.0
optimum>=1.13.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
